Add delete functionality

+++ app/src/main/java/ie/setu/appstore/models/AppModel.kt
@@ -4,5 +4,5 @@ import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class AppModel(var id: Long = 0,
+data class AppModel(var id: Int = 0,
                     var name: String = "") : Parcelable
